API/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.0.0
dj-rest-auth==4.0.1
dj-rest-auth-social==2.2.4
Django==3.2.20
django-allauth==0.54.0
django-braces==1.15.0
django-cors-headers==3.14.0
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
facebook-sdk==3.1.0
firebase-admin==6.2.0
google==3.0.0
google-auth==2.21.0
gunicorn==21.2.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0

